Basic UI updates for 2.3 (work in progress)
This commit brings the latest working draft from June 2011 to the current integration branch. The UI is still a work in progress and none of the changes we discussed in Cologne have been integrated yet.

The stylesheet changes in this commit should be used as a base for the ongoing development of the interface. The styles are combining Symphony 2.2 and the Symphony 3.0alpha concept release.

Related topics:

- #799: UI: Headline Markup
- #798: UI: Navigation
- #801: UI: System Messages