Enable use of return button in preferences view
By default, browsers select the first submit button in a form if the return button is
pressed inside an input field. But because extensions can add as many submit buttons
as they like and Symphony's is always last, this behaviour breaks.

This fix adds an invisible submit button to the beginning of that form to re-enable
that feature.