Simplify Pickable
This commit removes the new options `always` and `multiple` again at this really bloated the plugin. As a result of this, the select box will always be visible independent of the option count.

If an extension likes to hide the select box in this case this can be done with a simple `.hide()`.

This commit also changes how Pickable stores items internally in order to make it possible to add pickable items after the plugin has been initialised.