Fix upload field export
Make sure that the function does what it says: If I use the export mode `getFilename`, it should return the filename and not the filepath. If the filepath is needed, the mode needs a different name.

This is important for fields like SBL using the export to fetch the related field's values.