Fallback value for DrawerXMLElement
Since prepareAssociationsDrawerXMLElement is kind of
new, 99% of the fields do not implement it. Instead, they
rely on their prepareTableValue implementation to make
association drawer work. But, drawer now directly calls
preparePlainTextValue, some field were displaying nothing.
This fix actually calls prepareTableValue if the plainTextValue
is empty. This could be removed in future versions.