Delegate relation finding to fields
This proposed change make the fields reposible for
finding relations. This change takes methods that were
in content.publish.php and moves them to the Field class,
preserving all compatibility.
Reponsability is put in the right container and
Field implementors will now be able to defines their own mecanism
in order to create relationships.