Focus only when animation is over
This fix prevents a nasty bug where the handle (header) was something
clipped during the animation. This bug first appeared in 2.4.

The problem with this was that layout was invalidated when looking
for the first visible input.

A lot of ad-hic regression testing has been made in order to make sure that
nothing got broken.