Orderable optimization
Doing ui related things into event handlers such as mousemove or scroll
is a bad idea since the browser fowards all OS events it receives to javascript.

This creates a situation where multiple events are received during a single paint.
This is useless and inefficiant since the outcome of most handler is discarded by the last
received during the frame.

The correct way to do it is by calling request animation frame in order to delay execution
of the js and not block the UI thread.