More js optmization
1. Made collapse and expand actual function we can call: this removes
a lot of overhead that was created with usign events for internal communication.
All events were kept for backward compatibility and the global public api does not change.

2. Saving into localstorage now only occurs once, and it delayed: This added a lot
of extra work since local storage was accessed once by EVERY item in the collapsible.
Since the save checks every item, we ended up with O(n^2). With this, we still have
O(n) for events handling but we drop to O(1) for local storage access.

3. Change some jquery selector to be more precise.