Added Status 302 header in redirect()
I do not know why, but without it, my setup was returning a 500 code if the actual http status was something else (404 in my case).

This fix only makes sure that the 302 status (temporary redirect) will be returned to the client.
After all, a call to redirect must really mean that we want a redirect.

Note: I did not used the Page class helpers since this is 'utility' code and I did not wanted to couple utility code to the Page class.