Renamed `Gateway` class to `HTTPGateway`
As stated in #2229, the goal of this change is to make it more clear
what's this class is all about. I follow the principle that we already
have the EmailGateway (so prefixing "Gateway" with its related
"technology") and used UPPERCASE for HTTP since we already have the
`SMTPGateway` (uppercase) class name.