Adding a check to extension status before routing
This changes adds a new method on the ExtensionManager.

isInstalled($handle) will return true only if the extension is installed.

This change also calls this method in the Administration class in the getPageCallback method:
it checks if the extension is intalled before resolving the callback.

It acts as if the extension was not there when it is not enabled.

Fixes #2367