Adding a check to extension status before routing
This changes adds a new method on the ExtensionManager.

isInstalled($handle) will return true only if the extension is installed.

This change also calls this method in the Administration class in the getPageCallback method:
it checks if the extension is intalled before resolving the callback.

It acts as if the extension was not there when it is not enabled.

Fixes #2367
Picked from 066cbdcf57246515fb58e7b37f494e379f8124ed

Fix checking if an extension is enabled

RE: #2367

Picked from 92a0b82d5ce0d2d6717b0d939eb241b4c97aca84

Co-authored-by: Brendan Abbott <brendo@users.noreply.github.com>
Co-authored-by: Nicolas Brassard <nitriques@users.noreply.github.com>

Picked from 7406725ac8
Picked from 663b4e76bd