Disable GenericExceptionHandler before auth
Those changes makes sure we do not leak exception information before
doing the authentication. It also makes sure that exception raised in
the exception handler won't leak info if the GenericExceptionHandler is
disabled.

The same logic is also applied in the shutdown function.

Instead of masking the error like we did, the logic is applied only in
the render part of things, which should make it easier for us to debug.
It also removes the wrong HTTP status that we were returning (500
instead of 404)