Added Default Value plugin
This commit introduce a new jQuery plugin called Default Value, which
gives a input a source element to value its value from, when a specified
event occurs on this source element.

This behaviour is switched to off when the users focuses on the input
field and then blurs out, leaving a value in the input.

This commit also hooks this behaviours on the Data Source name input and
on the Event name input.

Closes #2511

Picked from 0e27ff517b