New class "field-value-readonly"
Adds a new class that allows for consistent styling of field values
that are displayed as plain text and are not editable (As seen in
extensions like „System Date Fields“, „Members“, „Member Last Visit“,
and maybe more).

See https://github.com/symphonycms/symphony-2/issues/2532 for details.

I’m planning to send PRs for the affected extensions after this has
published in the core.