Publish Filtering : Increase flexibility of comparison modes
The current markup of the „comparison mode“-selectboxes in the publish
filtering interface doesn’t allow for multiple modes that don’t have a
filter prefix (like „regexp:“, „greater than:“, etc.) as it uses the
prefix as unique identifier (option value).

The default comparison mode „is“ is expected as the only mode without a
filter prefix. But filter modes like „between“ („$x to $y“) as included
in Number Field (and possibly date fields too) don’t have a prefix
either and therefore won’t work with the current code.

This PR changes that behavior by shifting to the filter title as unique
identifier and adding the filter prefix as optional data-attribute
that’s only used for building the actual filtering-url-parameters.

Old Behavior:
- Filter title is used as option title
- Filter prefix is used as value-attribute

New Behavior:
- Filter title is used as option title
- Filter title is used as value-attribute
- Filter prefix is included as data-attribute