Use „data-comparison“ as unique identifier of the chosen comparison mode
Use the „data-comparison“-attribute as unique identifier for the chosen
comparison mode of a filtering instance - both in the selectbox and the
list of context help texts.

The „value“-attribute of the comparison mode selectbox now again is
used for serving the actual filter prefix used to construct the
filtering-parameters in the url.