Make sure user can delete the author (#2572)
This commits reuse the code from the edit action in order to validate if
the current user can delete the author.

Even if the Delete button is not present on the page, a request can be
crafted in order for an author to delete another.

This change also validates that the password of the current author is
checked before doing any un-undoable things.

Rebase of 8b20b8f

Picked from f7a5406dd1