Only request the needed schema in table view (#2568)
This changes uses the array returned by $section->fetchVisibleColumns() in order to create the element names array to pass to the entry manager. This yeilds a nice performance update when a section contains many (hundreads) of fields but only a few of them are needed for the publish table view.

Rebase of 34bd49f

Picked from 92f762bd95