Add data-attributes to publish filtering interface (#2577)
* Publish Filtering : Increase flexibility of comparison modes 6e764d1

The current markup of the „comparison mode“-selectboxes in the publish
filtering interface doesn’t allow for multiple modes that don’t have a
filter prefix (like „regexp:“, „greater than:“, etc.) as it uses the
prefix as unique identifier (option value).

The default comparison mode „is“ is expected as the only mode without a
filter prefix. But filter modes like „between“ („$x to $y“) as included
in Number Field (and possibly date fields too) don’t have a prefix
either and therefore won’t work with the current code.

This PR changes that behavior by shifting to the filter title as unique
identifier and adding the filter prefix as optional data-attribute
that’s only used for building the actual filtering-url-parameters.

Old Behavior:
- Filter title is used as option title
- Filter prefix is used as value-attribute

New Behavior:
- Filter title is used as option title
- Filter title is used as value-attribute
- Filter prefix is included as data-attribute

* Fix code indenting 4a7db7a

* Improve construction of comparison array ccc5e9e

* Fix code indenting e56bdde

* Use „data-comparison“ as unique identifier of the chosen comparison mode 	d37e0b8

Use the „data-comparison“-attribute as unique identifier for the chosen
comparison mode of a filtering instance - both in the selectbox and the
list of context help texts.

The „value“-attribute of the comparison mode selectbox now again is
used for serving the actual filter prefix used to construct the
filtering-parameters in the url.

* Remove leading whitespace from filter value input field ac61c11

Picked from 51ba3566a7