Make pickable hidden forms elements readonly
Not visible element should be made readonly, since the user can't edit
them anyways.

This fixes a problem where Chrome would auto-fill hidden input and the
value would be saved without the user ever seeing it.

We already tried to fix the problem usign a standard solution
(auto-complete="off", #1843 and #1841) but it does not work. We also tried a non-standard
solution (#2258) which was rejected.

This change only uses standard solution, even thought it relies on
javascript to make things works (it should not be a problem since the
backend now heavily relies on javascript)