Tag list field: accept an array of tags (#2602)
* Accept an array of tags 189b387

While the tag list field always returned string (parameter pool) and array values (XML) in the output, it did only accept a string of tags on post. This commit extends the field to accept an array of tags as well which is helpful, if you are building a tag list widget on the front-end and would like to post back to the system.

This change is supposed to be backwards compatible (it doesn't change the behaviour for posted string values).

* Fix whitespace 49a5200

Picked from 7e9c94e21f