PHP 7: Refactor around Exception handlers
PHP7 introduces a new interface, Throwable, which acts as the base class
for all Exception and Errors, unifying the concept at last. See http://php.net/manual/en/class.throwable.php for more info.

The changes proposed in this commit removes any class qualifier from the
method's signature, allowing both Exception and Throwable to be
handled.

Instead of checking the PHP version, this commit simply detect if a
Throwable class exists. If it doesn't (in PHP 5), it creates a
placeholder class, named Throwable, as an abstract sub class of
Exception. This make instance of check works and simplifies the code a
lot.

Try/catch clauses that want to catch *ALL* exception must be adapted.
See https://trowski.com/2015/06/24/throwable-exceptions-and-errors-in-php7/
This is not included in this commit since all not-cacheable Exception
will get picked up by our handlers and all custom Exception still
inherits from the Exception base class. The only thing that will not get
picked up would be Throwable that are not Exceptions, which never were
exception in PHP 5

Picked from 6f42cbd76f