Add the Content-ID header on attachments
This allows developer to reference attachments in the html body to allow
images to be viewed offline.

You can use it as follow:

```html
<img src="cid:name-of-the-file.ext" />
```

Where the name-of-the-file.ext part matches the actual file name of the
file. Notice the `cid:` prefix, which makes it work!

Picked from 2b75aa7b70