Sort settings of associated sections in drawer (#2608)
Respect sort settings of associated sections in associations drawer

Would make a lot more sense to me if the associations drawer would show entries of associated sections in the same order as the publish index table.

My quick and dirty hack seems to work fine so far, but not sure if this has any nasty side effects or if there's a more elegant solution to the problem (maybe #2553).

Also not sure if others agree with me on this matter in general, so feel free to dismiss this pull request if this is actually not a desired behavior.

On a side note, I guess it would be possible to provide this functionality as an extension, but the extension would have to duplicate most of the core code for the associations drawer. Might still be an option for bringing this to the 2.6.x branch as well, though.

5b37a73

Picked from 06d3da6c6a