Fix for MySQL 5.7 Strict mode
This is needed when sorting on a column that is not part of the projection.
In MySQL 5.7 strict mode, it is now required to add all columns in the ORDER BY clause in the SELECT's projection.

I could not find another way of fixing it, then adding a new parameter
which can inject the required missing piece in the select clause.

It's not a backward incompatible change on PHP 5.6, but can break
extensions on PHP 7.0+ since method signature must remain intact.
At least, the core is not broken anymore on MySQL 5.7 strict.

cc @brendo
cc @michael-e

Picked from 58bc57f2a4