Make sure we call stripslashes on read
And also call sanitize when creating the html.

When requests comes in, they are validated, then addslashes is called
before we write the php file. We still need to use the values from the
request, not call stripslashes on those since it's not needed.

This commit also adds a validation on resources names that prohibits the
use of \ in the name. Symphony already removed it, but the user was not
notified.

Re e30a18f8f09dca836e141bf126a26e565c9a2bc7
Re #2655

Picked from 18d24e086d