Make length paramter optional
The substr function does not require the length paramter to be always given.
This commit adds a default value (null) to the length paramter.
This commit also makes sure to not pass the null length paramter to substr, since it will always return an empty string. the mb_substr function is able to handle a null length (considering the length paramter to be unspecified when null).
That way, we make sure to always have the same behavior, regardless of the presence of the multybyte string module.