Fix unselectable element after a reorder
After a reorder of sections, elements were unselectable, due to a dirty
check that was failing. It's possible to get multiple
orderstop.orderable events without getting a orderstart.orderable event
beforehand. But, we already have all the information to update the
oldSorting value with the one we are about to save.

Fixes #2657

Picked from 5ab0a7b932