Prevents data loss on save (#2659)
We add a timestamp check before saving records, either in sections
(fields) or entries.

When saving a section, if the data changes between the render of the
page and the form post, it can lead to data loss if a field was added.
The SQL will crash because it assumes the fields did not change.

When saving and entry, the SQL was ok, but still, it can lead to data
loss.

At both places, the fix provided here makes sure we are editing the
lastest version of the record.

Rebased from 81438f64a5~1...0c8c76581e
Rebased from 60e930ee8f~1...1a931a4adc

Fix SQL error when saving a deleted field

This commit makes sure that a request to update a field will not fail if
the field as already been deleted by another user (or the same user in a
different tab/browser).

If the field is found to be non-existant, it is treated as a new field.

Fixes #2703
pick 1c84b2194b

Replace the added fieldset with an alert

@brendo's suggestion is WAY better then the current implementation,
which adds a new UI pattern in the system. Usign links in the alerts
makes it possible to offer the overwrite mode, without gluttering the
UI.

In order to make it possible to submit the form from the alert, a little
of pretty simple javascript was needed.

Rebased from 6f37eafdb6~1...c652b18d2a