Prevent acting on non existing tables
This change makes sure that the changes introduced in 4b5067f38 do not
cause a bug when the table does not exists. The old code (pre 2.7.0)
discarded all errors on delete, then checked if it had data to be
inserted. If not, the insert statement would be skipped.

The locking mechanism did add a lock statement before the delete, and
added logging all exceptions. That resulted in log "vomit".

This changes propose 2 ways of fixing this log problem:
1. It moves the empty data checks BEFORE the lock (so before the delete
as well).
2. It checks in the database if the table exist before doing anything.

4b5067f38 also un-lock the table before doing the insert. I think that
it should still maintain the lock until both delete and insert are done.

Fixes #2663

cc @animaux @michael-e