Replace url state with canonical url
I often see urls being shared with the various flags (saved, created) in them.
I also often see them while browsing the history.

Those ideally would not exists, but since we do a redirection on save, we need a way to still display the messages to the user.

The idea behind this commit is to hack the url usign javascript: if the page includes a link rel="canonical" tag in the head section, the javascript will replace the current state with the canonical url. It's also done only on full page load, to let extension read the value before replacing it.

This would not break the existing behaviour and could be removed when the redirection will be removed.

Rebased version of 3a4bed7