Make sure Authors emails are always validated
Authors' email are validated usign the 'email' validator.
Right now, validators are not being included in the validate function,
which rendered the validation pointless.

This commit make sure we have a defined 'email' validator.
If is does not, it will fallback to filter_var($email,
FILTER_VALIDATE_EMAIL).

Fixes #2632

Picked from 06ff493db9