Move General::buildPaginationElement >> DataSource
General is a already quite loaded. This method those not belongs there.

This commit replaces the implementation in General with a call to
DataSource. The General method is marked as deprecated. The General
method now calls the new implementation directly.

The new implementation uses an array to receive optional parameters via
an array, which makes the implementation a bit better.

Fixes #2670