Better fix for MySQL 5.7 Strict mode
58bc57 introduce a possible backward compatiblity change, since PHP 7
requires the child class to declare all optional paremeters, even if not
used. This would not be a breaking change 5.6.

Following was has been proposed in 58bc57, this commit removes the new
`$select` parameter in favor of a new method, `buildSortingSelectSQL()`.

This new method is called in the `EntryManager`, just after the call to
`buildSortingSQL`. That way, it is possible for fields that requires to
add columns into the project to do it, while not changing a public API.

Note that this is still a temporary fix that would need to be addressed,
but when a major version will come out.

Picked from 30294682f2