Add requiresTable perf optimization (#2678)
As suggested by @jonmifsud in #2664, this commit adds a new method in
the Field class, happily named requiresTable. Analogous to createTable,
this boolean flag tells Symphony to act on a table or not, reducing the
need to check for existing table when it is known to not exist.

Picked from d5f003d872