Add modification_author_id column
This is needed to properly display the name of the author that created
the conflict.

I first thought that the author_id would get updated on save, but it
turns our it does not: it's the creation author.

So in order to make it backward compatible a modification_author_id is
needed.

pick dff0d7f14e