Add unattended installation process (#2654)
This commit adds the ability to install Symphony without the need to
fill the install form manually. The only thing needed to implement it is
to fake POST data with a php file, namely manifest/unattend.php.

The unattended process can launch the installation process or simply
prefill the form. This is done with the 'action' array.

If something goes wrong, the form will be shown, with all the error
messages. Next form POSTs will override values coming from the
unattend.php file, giving you a quick way to fix the problems.

This feature was motivated with the need to be able to automatize the
install process without much change to the current code base.

Finally, a scaffolding version of a unattend.php file is provided in the
install/includes folder.

Re 	8e90d69