Refactor edit() and add() to make them identical (#2680)
4b5067f38c823b1534c0d57e72dd031465842e7c Added the logic need to locks
tables in order to perform the delete and insert in a more "isolated"
way. This is good, but it was only added on the edit() method. The add()
method is almost identical so a refactor introduced the saveFieldData()
method in order to offer locking mechanism to both methods.

But, it was noted that commit
04af0ffd8#diff-57eff69d6ab1319955c5a954c701ca33R126 silenced SQL error
for the DELETE statement, but 4b5067f38c823b1534c0d57e72dd031465842e7c
added LOCK and INSERT statements in the try/catch show.

All this time, add() was not modified: it stills throws exception if an
SQL error occurs.

With this refactor, add() gets all the protection and care we've put
into edit().

Rebased from 	d18d7cfb

Picked from 6c8d3688a7