Fix broken AND ( + ) filtering
Right now, usign a single + for the AND filter did not worked.
The + symbol we used is a space character in urls, where most of the
filters comes from.

So, the idea is to add a new split token, namely "+++" which makes
"filter 1 + filter 2" work.

This change remove code duplication in the publish page and in the section data source.

Fixes #2561