Fix broken filter after a sort
Sorting a column while a filter is active is removing the last character
of the filter from the URL after the sort is applied, if and only if, it
ends with one of those for chars: &amp;

I do not understand why both 7328c19a5 and 4631a4371 used
`trim($params['filters'], '&amp;');` to get rid of the leading '&amp;'
in the filter string.

This commits merges both mention commits into the same line, and
properly removes the leading &amp; chars.

Fixes #2465

Picked from b98f7d770b
Picked from 70175618c7