Fix not-regexp which leads to wrong results (#2695)
When filtering with not-regexp, it is required to do AND because we need
all columns to not match: having only one not matching is not enough.

Also, the if checked assumed that if the `regexp:` pattern was not
found, it would default to `not-regexp:`. Make sure that we check both.

Picked from f261fb35c2