Make sure we always have a date and time format (#2705)
If, somehow, the date_format or time_format config is missing, Symphony
stops displaying dates. This is problematic as we rely on date
formatting a lot.

This change adds a default values for date_format and time_format
directly in the DateTimeObj class. It also makes sure we have a valid
timezone. If not, the date.timezone php.ini value is used. If it is
still empty, we default to UTC.

This change made is possible to remove Configuration()->get() calls and
replace them with __SYM_DATE_FORMAT__ or __SYM_TIME_FORMAT__ constant
when possible, or uses DateTimeObj::getSetting() which shoulw always
return a value.

Finally, this change also adds page alerts when a missing date_format or
time_format is empty.

Fixes #2704

Picked from 4efc979
Picked from 3cdf4f3d10