Make sure buildSortingSelectSQL() does not crash
The default implementation of `buildSortingSelectSQL()` is well suited
for Fields that do not provide a custom `buildSortingSQL()`
implementation. But this can make Fields with a custom `buildSortingSQL()`
implementation without a `buildSortingSelectSQL()` crash if they do not
have the 'value' column.

This commits adds an extra SQL query to first check that the colunm
exist. For performance reason, this result is memoized.

Fixes #2718