Restore the 2.6.x field behavior for empty arrays
Symphony always treated empty arrays as being null values, i.e. deleting
old data without doing an insert statement afterwards.

The refactor made in 2.7.x used the code for new as the base, instead of
edit. new never had to deal with old data and simply did not produce a
DELETE statement, but edit did.

I think it is safe to change both, to always issue a DELETE statement
and produce an INSERT when needed.

Fixes #2733