Fix escaping of invalid XML (#2750)
- Fixes an issue with unencoded special characters in an input field that expects valid XML

Picked from b5280f6e7f

- Do not change the value in the $fields array

In case where the XML is not valid, we do not want to escape it.
Valid XML gets saved AS IS, while invalid XML is not saved, but returned
to the browser also AS IS.

The only needed escapes is when we set the textbox's value.

Picked from 409f000c92

Fix #2748
Re #2749
Re #2750