Remove FILTER_SANITIZE_STRING filtering
FILTER_SANITIZE_STRING is too agressive. FILTER_FLAG_STRIP_LOW is
softer since we do not want to keep control chars.

BUT: doing so makes the platform vulnerable to XSS. This is required
because setting can contains specials characters (like > and <) or html.

The responsability to properly escape data is then being pass down to
extensions.