Fix error if extension not found on boot
Before symphony can be properly booted, the extension manager is created
in order to allow extensions to override the launcher.

If there is an extension missing, the user gets a non helpfull message
telling him that Symphony did not boot.

This commit makes sure the user is notified about the error even if a
nice error template can not be displayed.