Fix field duplication without field settings
Some fields do not create any records in their settings table because they do not
implement a proper `Field::commit()` method.
The base implementation of the commit function only deals with the "core"
`tbl_fields` table.
The problem with this approach is that it can lead to data corruption when
saving a field that got deleted by another user.
The only way a field can live without a commit method is if it does not store any
settings at all.
But current version of Symphony assume that the `tbl_fields_$handle` table exists
with at least a `id` and `field_id` column, so field are required to at least create
the table to make their field work without SQL errors from the core.

Fixes #2736