Make DS params in XML optional
This commits add a new parameter in data sources to control if the
output parameters should be added in the xml output.

By default, data sources that do not have the dsParamPARAMXML property
set will include their output parameters in the XML, since this is their
current behavior. If the dsParamPARAMXML exists on the object, it only
needs to be different from 'yes' to be excluded from the XML.

Fixes #784