Better error handling and disabled display of exception data (#2597)
* Fix doc: E_NOTICE are loggued c8ff3dc

The doc stated that E_NOTICE were ignored while the code only ignores
E_STRICT (see line 423)

* Make renderHtml protected ebc5ac7

This method should never get called directly.
Nothing ever calls it in the current base code.

* Disable GenericExceptionHandler before auth d53b221

Those changes makes sure we do not leak exception information before
doing the authentication. It also makes sure that exception raised in
the exception handler won't leak info if the GenericExceptionHandler is
disabled.

The same logic is also applied in the shutdown function.

Instead of masking the error like we did, the logic is applied only in
the render part of things, which should make it easier for us to debug.
It also removes the wrong HTTP status that we were returning (500
instead of 404)

Picked from dbcf209

* Add GenericExceptionHandler::$logDisabled dec526d

This implementation copies what was already present in the
GenericErrorHandler, i.e. skip logging if disabled = true.

This also prevents errors caught in GenericErrorHandler to be also
loggued a second time in the GenericExceptionHandler::handler method.

* Check GenericExceptionHandler::$enable in subclass 20f37aa

Both SymphonyErrorPageHandler and DatabaseExceptionHandler needed care:

SymphonyErrorPageHandler:
- throwCustomError must not touch GenericExceptionHandler::$enabled
  since it allows data leakage
- defaults the diabled and not template cases to
  GenericExceptionHandler::render

DatabaseExceptionHandler:
- Filter out data if the Handler is not enabled. This handler was
  already getting the right template since it uses the same as the
parent.

GenericExceptionHandler:
- refactored the headers part to share code in SymphonyErrorPageHandler

* Made handlers final 8d84a7d

Both the GenericErrorHandler and the GenericExceptionHandler's handler
function should not be overriden.

* FrontendPageNotFoundExceptionHandler bypass 809f927

FrontendPageNotFoundExceptionHandler can bypass the new anti-leak
system since it's a legitimate message.

Picked from 3da6eeb9a8