Redistribute responsabilities between Page/Process
We used to store the xml and xsl values both in the XSLTPage AND in the
XSLTProcess. This is redundant and totally not needed.
This change makes only the Page responsible to hold the xml and xslt
data, which are simply passed when the XSLTProcess needs to do it job.

The only reason why the Process would need to hold the data is when
there is an error. For this purpose, the "context" is managed manually
instead of exposing it publicly.

On the other hand, the Page was holding data belonging to the Process:
params and functions. Since the page does not use the parameters anymore
and it does not register any functions, this data has been moved to the
process. Upon instanciation, the Page gets a process, so developers can
just forward their Page calss to Page->Proc.

The end goal was to not expose any XSLT parameters or function in the
XSLTPage, which is now acheived.