Fix bugs with fetch() unsafe where
Fields always prefix their where clause with ' AND ' so we need to make
sure that we add a void WHERE clause if it is the first to be added.

We also make sure to catch the '`t1`.' table prefix variant.

Picked from 0e6cb75fbf
Picked from d5c524a586
Picked from 77fe391223