Add the Content-ID header on attachments (#2617)
* Add the Content-ID header on attachments d7abb0ecceba646eb235cdaa6a703b4cdb639ce3

This allows developer to reference attachments in the html body to allow
images to be viewed offline.

You can use it as follow:

```html
<img src="cid:name-of-the-file.ext" />
```

Where the name-of-the-file.ext part matches the actual file name of the
file. Notice the `cid:` prefix, which makes it work!

Picked from 2b75aa7b70

* Add a flag to enable Content-ID header field 5483f71e075df0458798c2b0c1e2b13bb499fb37

* Let the developer controls the disposition value c3a6a270771508461019aa5a3d760029c39dd054

* Let the developer controls the content-id value 87103096d5b06316db94d03f82eaa8c7aa0714d6

* Remove null keys from array and add isset() 24786e17c82426b14ca50e428838f1054a6a3974

Also remove the old undocumented format

* Add missing commas 5c759921ae4b8b74f88672542f7436b0e32cf9e0

Picked from 895d7cb494