Add EntryQueryAuthorAdapter API
This new API removes a responsability in the Field class, i.e.
generating the SQL needed for filtering and sorting.

It adds a new method on the EntryQuery class, filter() and adapts the
the sort() method to delegate the filter/sort logic in the Fields'
EntryQueryAuthorAdapter object.

EntryQueryAuthorAdapter should support all filters required to make the
section data sources work, while offering the data source filtering
capibilities in the code also.

Picked from 7eb1f515d2