Add EQFA for Date Field
The new implementation supports 'regexp:', 'sql:' and 'not:' filtering modes.
If not, it will try to parse the filter with the DateRangeParser class.

The old implementation filter parsing as been moved into the
DateRangePaser and the usage remains the same. The only addition is the
'from ' prefix not allowed in closed ranges (from X to Y).

The old implementation did not had the 'not:' filtering mode that was in
this one.

Picked from e26cf0d3d9
Picked from a74a46c33f
Picked from 8cd0b1ed44