Refactor factory methods
Passing the $root_element name as a parameter is error prone, and hard
to do on unknown sources. This commit deprecates both existing factory
methods convertFromXMLString() and convertFromDOMDocument() and adds
fromXMLString() and fromDOMDocument().

fromDOMDocument() does not uses DOMXPath: we do not need it to traverse
the whole tree. Also, is makes sure to keep the root node name and its
attributes, which both were missing from the previous implementation.