Fix broken non-breaking space replacement
475c198acd (#2631) added support for non-breaking space, but github
editor converted it to a regular space.

This breaks with the new way we create handles so we need to convert
non-breaking space into an actual breaking (regular) space to allow
custom delimiter to work.

The commit also ports all createHandle tests from GeneralTest which
should continue to work + adds transliterations tests