Add convertValuesToFilters remove fetchIDfromValue
This makes sure the feature is documented in the public API.
The name has been change to reflect better the usage.

Fixes #2171

It also fixes the broken prepropulation when the new EQFA API is used.
Finally, it checks to see if the field supports the feature before
adding anything into the query strings.