Fix missing calendar suggestion with field errors
The calendar requires that the ul be next in the DOM to the input.
The current error handling code would wrap the label too soon, resulting
in the calendar being added to the wrong XMLElement.

This commits does the error handling at then end of the function, to
make sure the field is 100% build before wrapping it.

The error mode would also bleed its color css property into the
calendar, making the day number invisible (white on white). The commit
also includes setting a color a the .day class

Fixes #2857