Do no escape forward slashes (/) in env json
The json_encode function follows the standard at https://www.json.org/
which specifies to espace forward slashes. But doing so, makes all url
looks to be broken, because they are escaped.

In this particular case, since the json is outputed in a html page,
which makes it only consumable via javascript, it is safe to not espace
forward slashes, and allow reverse proxy to rewrite the urls, fixing a
bunch of problems!