Make sorting entries deterministic (#2872)
MySQL does not offer deterministic order bys. This leads to problems
where the user could see records move in the list or see a record that
was displayed in the previous page, which it not nice UX.

It is up to the programmer to create a deterministic sort. When the
fields contains mostly random data, the problem is not visible. But when
there are multiple records with the same value, the problem becomes
apparent: our order by clause is not unique.

The simplest fix I could find was to add the entry id as the last order
by clause. This clause needs to follow the same sort direction, to make
sure the last item becomes the first one when the order is flipped.

Fixes #2856