Validate section id upon entry edit
This commit adds a validation on the section in the url version the
entry's section id.

Without this patch, it is possible to load entries in any section.
If the user hit save, it wipes all data in the entry and does not give
any errors back to the user.

Even if Symphony does not generate such invalid links, it must not allow users
to wipe data like this.