(fix) Check if INTL_IDNA_VARIANT_UTS46
It turns out that this value not only depends on the php version, but
also on the version of the icu lib used to compile/link php.

php.net states
> INTL_IDNA_VARIANT_UTS46 only available as of ICU 4.6

This commit make it work on older versions: We need to at least
overwrite the default value of the parameter which as been deprecated in
php 7.2 and has been replaced in php 7.4.